토큰 기반 댓글 작성 기능 업데이트 및 각 api url 변경

+++ routes/posts.js
@@ -1,11 +1,85 @@
 const express = require("express");
 const router = express.Router();
-const postSchema = require('../schemas/post')
-//게시글 작성 API
-router.post('/posts', async(req, res) => {
-    const { user, password, title, content} = req.body
+const postSchema = require("../schemas/post")
+const commentSchema = require("../schemas/comment")
+const authMiddleware = require("../middlewares/auth_middleware")
 
-    await postSchema.create({ user, password, title, content});
+//댓글 작성api
+router.post('/posts/:postId/comments', authMiddleware,  async(req, res) => {
+    const { userId } = res.locals.user;
+    const { postId } = req.params;
+    const { comment } = req.body
+    if(!postId || !comment){
+        res.status(400).json({
+            "message": "데이터 형식이 올바르지 않습니다"
+        })
+        return;
+    }
+    
+    await commentSchema.create({userId, comment, postId});
+
+    res.status(201).json({"message" : "댓글을 생성하였습니다."})
+});
+
+// 댓글 목록 조회 api
+
+router.get('/posts/:postId/comments', async(req, res) => {
+    const { postId } = req.params;
+    const comments = await commentSchema.find({postId:postId})
+    const results = comments.map((detail) => {
+        return {
+            commentId: detail._id,
+            user: detail.user,
+            content: detail.content, 
+            //작성시간 확인필요..                        
+        }
+    })
+    if(comments){
+        res.json({data: results});
+    }
+});
+
+//댓글 수정 api
+router.put('/posts/:postsId/comments/:_commentId', async(req, res) => {
+    const { _commentId } = req.params;
+    const { password, content} = req.body
+
+    const changecomment = await commentSchema.find({ _id : _commentId })
+
+    if (!changecomment) {
+        res.status(404).json({
+            "message" : "댓글 조회에 실패하였습니다."
+        })
+        return;
+      };
+
+    await commentSchema.updateOne(
+        { _id: _commentId },
+        { $set: { password: password, content: content } }
+    )
+    res.status(200).json({"message": "댓글을 수정하였습니다"});
+});
+
+//댓글 삭제 api
+router.delete('/posts/:postId/comments/:_commentId', async(req, res)=> {
+    const { _commentId } = req.params;
+    const { password } = req.body;
+
+    const existcomment= await commentSchema.findOne({ _id : _commentId });
+
+    if(existcomment.password === password){
+        await commentSchema.deleteOne({ _id : _commentId });
+    }
+    res.json({"message" : "댓글을 삭제하였습니다."})
+});
+
+//게시글 작성 API
+router.post('/posts',authMiddleware, async(req, res) => {
+    const { userId } = res.locals.user;
+    console.log(userId);
+    const { title, content} = req.body
+
+    await postSchema.create({ userId, title, content});
 
     res.status(200).json({
         "message" : "게시글을 생성하였습니다."
@@ -45,16 +119,17 @@ router.get('/posts/:_postsId', async(req, res) => {
 
 //게시글 수정 API
 
-router.put('/posts/:_postsId', async(req, res) => {
+router.put('/posts/:_postsId', authMiddleware, async(req, res) => {
+    const { userId } = res.locals.user;
     const { _postsId } = req.params;
     const { password, title, content} = req.body
 
-    const post = await postSchema.find({ postId : _postsId })
+    const post = await postSchema.find({ userId, postId : _postsId })
 
     if (post.length) {
         await postSchema.updateOne(
-          { postId: _postsId },
-          { $set: { password: password, title: title, content: content } }
+          { userId, postId: _postsId },
+          { $set: { title: title, content: content } }
         )
       };
     
@@ -62,11 +137,12 @@ router.put('/posts/:_postsId', async(req, res) => {
 });
 
 //게시글 삭제 API
-router.delete('/posts/:_postsId', async(req, res)=> {
+router.delete('/posts/:_postsId', authMiddleware, async(req, res)=> {
+    const { userId } = res.locals.user;
     const { _postsId } = req.params;
-    const existposts = await postSchema.find({ _postsId });
+    const existposts = await postSchema.find({ userId, _postsId });
     if(existposts.length){
-        await postSchema.deleteOne({ _postsId });
+        await postSchema.deleteOne({ userId, _postsId });
     }
     res.json({"message" : "게시글을 삭제하였습니다."})
 });
